encode gas limit
